requirements.txt
asgiref==3.8.1
boto3==1.34.138
botocore==1.34.138
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
jmespath==1.0.1
mysqlclient==2.2.4
packaging==24.1
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
psycopg2==2.9.10
